Add Android API check task: compileAndroid
Adds a task named 'compileAndroid' which will compile miniDNS with the
Android runtime jar set as bootclasspath. This requires the environment
variable ANDROID_HOME to be set.

Running this test shows that miniDNS already uses some APIs that are not
available on all Android versions:

- Arrays.copyOfRange (Since API 9)
- IDN (Since API 9)
- AutoCloseable, in the try-with-resources statement (Since API 19)

So the minimum required Android API level of miniDNS is API 9, since the
try-with-resources statement can be easily changed back to old
try-finally-ifNotNull idiom.