Filter out empty values in AndroidUsingExec
Sometimes 'getprop' does return *.dns(1|2|3|4) entries with the empty
string as value. For example

[net.r_rmnet_data0.dns1]: []
[net.r_rmnet_data0.dns2]: []

The successive InetAddress.getByName(String) will then return a
loopback address, but there is usually no DNS resolver listening for
requests on that address. This causes the DNS lookup to take longer
than necessary, due to the timeouts of queries send to that address.

Thanks to Grigory Fedorov for providing information which helped to
solve this issue.